Reuse forward ref on all component types
