+++ package.json
@@ -11,7 +11,7 @@
   "scripts": {
     "test": "node node_modules/jest/bin/jest.js",
     "release": "wnpm-release --no-shrinkwrap",
-    "build": "babel index.js --out-dir dist --source-maps"
+    "build": "rollup -c"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
@@ -25,8 +25,10 @@
     "babel-preset-stage-3": "^6.24.1",
     "carmi": "^0.2.3",
     "jest": "^23.6.0",
+    "prop-types": "^15.7.2",
     "react": "^16.6.1",
     "react-test-renderer": "^16.6.1",
+    "rollup": "^1.26.0",
     "wnpm-ci": "latest"
   },
   "peerDependencies": {
