+++ rollup.config.js
@@ -1,0 +1,7 @@
+export default {
+    input: 'src/index.js',
+    output: {
+        file: 'dist/index.js',
+        format: 'cjs'
+    }
+};
